Move Python extension module initialization code to modmesh/python/
