+++ cpp/modmesh/buffer/pymod/TypeBroadcast.hpp
@@ -28,7 +28,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <pybind11/pybind11.h>
+#include <pybind11/pybind11.h> // Must be the first include.
 #include <pybind11/numpy.h>
 #include <modmesh/buffer/SimpleArray.hpp>
